refactor(game): extract game-setup cluster into GameSetupMixin (#1271)
Pull the create/reset/end/rematch game-setup cluster out of the state.py
God-Object into a new GameSetupMixin (game/state_setup.py). Pure code-move,
behavior- and API-preserving.

Moved: create_game, _reset_game_internals, end_game, rematch_game and the
Apple-Music _detect_storefront helper (used only by the two builders).

The _set_phase transition chokepoint, _apply_config/_default_config, the
owned managers and clear_all_sessions all stay on GameState; the mixin
references them via self. GamePhase is imported lazily inside the methods
that need it to avoid a circular import back into state.py. The
self.difficulty: str annotation (the #1271 has-type fix) moves with
create_game, so game/service.py's gated read stays resolvable.

state.py: 1360 -> 1078 lines (-282). Mixin added alphabetically to the
GameState base list with a docstring paragraph.
